fix(zsh): load macos plugin after os-agnosting plugins
the macos plugin runs `brew shellenv` which puts brew's path into
`${PATH}`. in our `path` plugin we put the go/cargo etc paths in there.
if we run the macos plugin after, it ends up earlier in the path, since
it prepends itself. that means that the go/cargo binaries don't get used
if the same name is available via brew. i want the locally installed
ones to be used. swapping the order fixes that.