feat: dynamically choose bat theme, fix ansible linting, sync directories
The original goal was to use Catppuccin themes for `bat`. For that I
needed to write a wrapper script and break out `is-dark-mode` into
somewhere on the PATH.

While I was modifying the Ansible YAML to install the new files, I
noticed that `prettier` and `ansible-ls` were fighting over the
indentation. So I had to figure out how to let `prettier` take charge of
that and stop the language server handling YAML formatting.

Finally, removing the old functions in favour of scripts made me realise
that we should use the `synchronize` module instead of `copy` for
setting up the configurations. With `delete: yes` this handles deletion,
so we don't end up with files left behind on the system itself if we
drop them from the repo.

A bonus which I should have committed separately. Set up `MANPAGER` to
colourise manpage output with `bat` as the pager.