fix: use full path to bat in user role, select dark/light themes natively
Turns out that `bat` supports env vars for dark/light themes, so we can
use that instead of our wrapper script.

Also, when the playbook is run in a single run, the `bat` binary is not
yet in the path, so we need to use the full path to it.