feat: support linux servers, add `cripps`
This is my VPS and I want much of the same configuration as my desktop.
add support for this, split the playbook up into desktop vs non-desktop.
Additionally, this is a Debian system rather than an ubuntu one, so
update family support to target all Debian derivatives.

This surfaced a couple of bugs

- Language package managers won't be in the path in the same run where
  they're installed. In a few places we want to call them - make those
  use the absolute paths.
- `pipx` installs itself to `~/.local/bin/` by default. One of the roles
  syncs files into there with `delete: true`. So they're fighting over
  the directory. Fix this by putting pipx in its own dir and putting
  that in the path.