feat(nvim): configure diffview & use for git mergetool
I wanted a nice way to do three-way merges when resolving git conflicts.
`diffview` seems like a good tool to do this, so configure it. While
we're there, fix up lazygit to work with nvim's shortcut. At least with
LazyVim it expects a config file to be there, so create one.