fix(nvim-lspconfig): fix eslint autofixing
The method LazyVim is using to configure autofixing for eslint is not
working. It has a < 0.10 and a >= 0.10 codepath. In the latter the LSP
support is used, and in the former `:ESlintFixAll` is called.

The [version detection used upstream][code] doesn't work for 0.11. Seems
that `require("vim.lsp._dynamic")` is not a reliable test. So on 0.11 we
enter the `:ESlintFixAll` codepath, but this doesn't work. We can fix
this up by directly checking the Neovim version instead of looking at an
internal API which has apparently changed.

But hmm, why did the fallback not work? It feels like it should have.
After a bit of poking, turns out that it's separate breakage! There's an
optimisation present to avoid running `:ESlintFixAll` when there is no
need - when there aren't any diagnostics - but it doesn't work. It
always reports that there are zero diagnostics, so we skip running
ESLint. As far as I can tell it's because the language server for ESLint
[now uses][eslint] [pull diagnostics][pull].

So what happened here was that first one thing broke (finding the
diagnostics, preventing the fallback path from running `:ESlintFixAll`).
But we were correctly not running the fallback path, so didn't notice.
But then then later when 0.11.0 of Neovim came out we started entering
that codepath again due to the other breakage, and then hit this latent
bug, Both auto-formatting codepaths broke and so there was no more
auto-formatting.

There's a fix for the latter too. `vim.lsp.diagnostic.get_namespace()`
has a second boolean parameter called `is_pull`. If this is `true` (the
default is `false`) then we get the `pull` diagnostics, This is what
modern ESLint language servers are using, so the test starts working
again. For backwards compatibility we can easily check for both, so we
don't break any people who are on old versions and so are still working
today.

Good times. Both paths work again.

[code]: https://github.com/LazyVim/LazyVim/blob/ec5981dfb1222c3bf246d9bcaa713d5cfa486fbd/lua/lazyvim/plugins/extras/linting/eslint.lua#L41
[eslint]: https://github.com/microsoft/vscode-eslint/commit/2c3d198cc648005f8cf5ca64476b8f2da8fa0205
[pull]: https://atlee.ca/posts/pull-diagnostic-support-for-neovim/