feat(nvim): add a `join-paragraphs` plugin
This is a `:JoinParagraphs` command/keybinding to ... join paragraphs
... What it does is run `Join`, but select the surrounding paragraph
automatically and handle visual / normal mode, and ranges too.

There's a `:Rj` command, which is a mnemonic for "read and join": run
`:read` and then `:JoinParagraphs` on the result.

Quite complicated, since there are some (literal) edge cases. We need to
wrap each paragraph separately. So we have to try to find those (by
looking for empty lines). But then this breaks at the end of the buffer,
since the last paragraph is not followed by an empty line. So that needs
special handling - we add a temporary new line there. But then we have
to remove it. Oh dear. You get the idea!

I should probably publish this separately!