feat: add biome extra, make it not conflict with prettier
Giving this a go, but it needs some integration. It can't run at the
same time as Prettier, or they conflict. Additionally, it surfaced that
none-ls was being prioritised over conform. Ideally we'd use conform
where it supports a formatter, and none-ls as a fallback.

We configure biome to use `biome-check`, which formats and lints more
stuff than the default `biome` formatter.