feat(nvim): update, switch copilot plugin to blink-copilot, add avante and claudecode AI plugins
Upstream is doing a lot of work integrating AI plugins - keep up to
date. We're using a new copilot plugin, blink-copilot, which is better
integrated with blink.cmp. I've not been super happy with the blink
setup up until now, so switch to the their "super-tab" keymap preset
which is more like VS Code's behaviour. (tab rather than enter to
accept, for example.)