chore(jsonnet): switch to native LSP config
In 0.11, Neovim introduced a new way of configuring LSP servers via
`vim.lsp.enable()`. Here we update to that.

To be honest, I did this a few weeks ago. I don't _think_ I did this
just for fun, but because something was broken, but I forgot what that
was or even if that's true. Ah well.